refactor: domain, dto 패키지 분
