+++ src/main/java/roomescape/console/dao/MemoryReservationTimeDao.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicLong;
 import roomescape.core.dao.ReservationTimeDao;
-import roomescape.web.domain.ReservationTime;
+import roomescape.core.domain.ReservationTime;
 
 public class MemoryReservationTimeDao implements ReservationTimeDao {
