+++ src/main/java/roomescape/console/view/OutputView.java
@@ -1,8 +1,8 @@
 package roomescape.console.view;
 
 import java.util.List;
-import roomescape.web.dto.response.ReservationResponse;
-import roomescape.web.dto.response.ReservationTimeResponse;
+import roomescape.core.dto.response.ReservationResponse;
+import roomescape.core.dto.response.ReservationTimeResponse;
 
 public class OutputView {
