+++ src/main/java/roomescape/core/domain/ReservationTime.java
@@ -1,4 +1,4 @@
-package roomescape.web.domain;
+package roomescape.core.domain;
 
 import java.time.LocalTime;
