+++ src/main/java/roomescape/core/service/ReservationService.java
@@ -4,10 +4,10 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.core.dao.ReservationDao;
 import roomescape.core.dao.ReservationTimeDao;
-import roomescape.web.domain.Reservation;
-import roomescape.web.domain.ReservationTime;
-import roomescape.web.dto.request.ReservationRequest;
-import roomescape.web.dto.response.ReservationResponse;
+import roomescape.core.domain.Reservation;
+import roomescape.core.domain.ReservationTime;
+import roomescape.core.dto.request.ReservationRequest;
+import roomescape.core.dto.response.ReservationResponse;
 
 @Service
 public class ReservationService {
