+++ src/main/java/roomescape/core/service/ReservationTimeService.java
@@ -3,9 +3,9 @@ package roomescape.core.service;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.core.dao.ReservationTimeDao;
-import roomescape.web.domain.ReservationTime;
-import roomescape.web.dto.request.ReservationTimeRequest;
-import roomescape.web.dto.response.ReservationTimeResponse;
+import roomescape.core.domain.ReservationTime;
+import roomescape.core.dto.request.ReservationTimeRequest;
+import roomescape.core.dto.response.ReservationTimeResponse;
 
 @Service
 public class ReservationTimeService {
