+++ src/main/java/roomescape/web/dao/JdbcReservationTimeDao.java
@@ -10,7 +10,7 @@ import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.core.dao.ReservationTimeDao;
-import roomescape.web.domain.ReservationTime;
+import roomescape.core.domain.ReservationTime;
 
 @Repository
 public class JdbcReservationTimeDao implements ReservationTimeDao {
