+++ src/test/java/roomescape/reservation/dao/JdbcReservationDaoTest.java
@@ -14,8 +14,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import roomescape.web.dao.JdbcReservationDao;
-import roomescape.web.domain.Reservation;
-import roomescape.web.domain.ReservationTime;
+import roomescape.core.domain.Reservation;
+import roomescape.core.domain.ReservationTime;
 
 @JdbcTest
 class JdbcReservationDaoTest {
