+++ src/test/java/roomescape/reservation/service/ReservationServiceTest.java
@@ -14,9 +14,9 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
 import roomescape.core.service.ReservationService;
 import roomescape.core.service.ReservationTimeService;
-import roomescape.web.dto.request.ReservationRequest;
-import roomescape.web.dto.request.ReservationTimeRequest;
-import roomescape.web.dto.response.ReservationResponse;
+import roomescape.core.dto.request.ReservationRequest;
+import roomescape.core.dto.request.ReservationTimeRequest;
+import roomescape.core.dto.response.ReservationResponse;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
